move Include to bokeh.core.property.include
